+++ src/Operation/Extensions.cs
@@ -158,5 +158,135 @@ namespace System
             return Throw(operation, message);
         }
 
+
+
+        public static Operation<T> Fold<T>(this IEnumerable<Operation<T>> operations, Func<T, T, T> accumulate)
+        {
+            using (IEnumerator<Operation<T>> e = operations.GetEnumerator())
+            {
+
+                //Make Sure Sequence isn't empty
+                if (!e.MoveNext())
+                {
+                    var ex = new InvalidOperationException("Sequence contains no Elements");
+                    return new Operation<T>(ex)
+                    {
+                        Message = ex.Message,
+                    };
+                }
+
+                List<Operation> badOperations = new List<Operation>();
+
+                //Process First Element
+                var result = default(T);
+
+                if (e.Current.Succeeded)
+                {
+                    result = e.Current.Result;
+                }
+                else
+                {
+                    badOperations.Add(e.Current);
+                }
+
+                //Process the Rest
+                while (e.MoveNext()) 
+                {
+                    if (e.Current.Succeeded)
+                    {
+                        result = accumulate(result, e.Current.Result);
+                    }
+                    else
+                    {
+                        badOperations.Add(e.Current);
+                    }
+                }
+
+
+                //If there are any bad operations
+                if (badOperations.Any())
+                {
+                    string messages = null;
+                    for (int i = 0; i < badOperations.Count; i++)
+                    {
+                        //Get Innermost Exception
+                        var badOperation = badOperations[i];
+
+                        //If this this the First Exception then Assign it Directly
+                        if (i == 0)
+                        {
+                            messages = badOperation.Message;
+                        }
+                        else
+                        {
+                            //Othewise Concat the Message
+                            messages = messages + ", " + badOperation.Message;
+                        }
+                    }
+
+                    //Aggregate Exceptions
+                    var aggregate = new AggregateException(badOperations.Select(o => o.GetException()));
+
+                    return new Operation<T>(aggregate)
+                    {
+                        Message = messages,
+                        Succeeded = false
+                    };
+                }
+                else
+                {
+                    return new Operation<T>()
+                    {
+                        Succeeded = true,
+                        Result = result
+                    };
+                }
+            }
+        }
+
+
+        public static Operation Fold(this IEnumerable<Operation> operations)
+        {
+            using (var e = operations.GetEnumerator())
+            {
+                if (!e.MoveNext())
+                {
+                    var ex = new InvalidOperationException("Sequence contains no Elements");
+                    return new Operation(ex)
+                    {
+                        Message = ex.Message,
+                    };
+                }
+
+                List<Operation> badOperations = new List<Operation>();
+                do
+                {
+                    //If the Current Operation did not succeed, add it
+                    if (!e.Current.Succeeded)
+                    {
+                        badOperations.Add(e.Current);
+                    };
+                }
+                while (e.MoveNext());
+
+
+                //If there were any bad operations
+                if (badOperations.Any())
+                {
+                    return new Operation
+                    {
+                        Message = badOperations.Select(o => o.Message).Aggregate((ag, m) => ag + ", " + m),
+                        Succeeded = false
+                    };
+                }
+                else
+                {
+                    return new Operation()
+                    {
+                        Succeeded = true,
+                    };
+                }
+            }
+        }
     }
 }